ref: Move sampler out of integrations/opentelemetry
