+++ sentry_sdk/integrations/opentelemetry/span_processor.py
@@ -13,6 +13,7 @@ from opentelemetry.sdk.trace import Span, ReadableSpan, SpanProcessor
 
 import sentry_sdk
 from sentry_sdk.consts import SPANDATA
+from sentry_sdk.sampler import create_sampling_context
 from sentry_sdk.tracing import DEFAULT_SPAN_ORIGIN
 from sentry_sdk.utils import get_current_thread_meta
 from sentry_sdk.profiler.continuous_profiler import (
@@ -21,7 +22,6 @@ from sentry_sdk.profiler.continuous_profiler import (
     try_profile_lifecycle_trace_start,
 )
 from sentry_sdk.profiler.transaction_profiler import Profile
-from sentry_sdk.integrations.opentelemetry.sampler import create_sampling_context
 from sentry_sdk.integrations.opentelemetry.utils import (
     is_sentry_span,
     convert_from_otel_timestamp,
